+++ packages/capabilities/scheduling/src/index.ts
@@ -1,3 +1,6 @@
 export * from "./model";
+export * from "./calendar";
+export * from "./cpm";
+export * from "./baseline";
 export * from "./service";
 export * from "./manifest";